backend/requirements.txt
fastapi
uvicorn
python-docx
docx2pdf
python-multipart
